travel-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
openai
langchain-core
langgraph
typing-extensions
langsmith
psycopg2-binary
